+++ NEWS.md
@@ -1,3 +1,11 @@
+# scRepertoire VERSION 2.5.1
+
+## UNDERLYING CHANGES
+* Introduced pairwise calculations to ```StartracDiversity()```
+* Internal function conversion for ```clonalSizeDistribution()``` - remove 
+cubature, truncdist and VGAM from dependencies.
+* Increased speed of ```clonalSizeDistribution()```
+
 # scRepertoire VERSION 2.5.0
 
 ## UNDERLYING CHANGES